coffee_shop/requirements.txt

Django==5.1
pillow==10.4.0
djangorestframework==3.15.2


psycopg2-binary==2.9.9

django-environ==0.11.2
